Add standalone fingerprinting utility
The fingerprint/compare pipeline is useful for retrospective analysis,
but not as useful when we're making dynamic changes to the proxy servers
in realtime to make them more compatible with Chromium. Refactor the comparison utility to accept a generic list of files for comparison (allowing multiple proxies to be compared at the same time) and add a new CLI to allow for packet sniffing + result comparison in one stage.