Fix Chrome certificate validation
We were running into an issue specifically with the `ChromeRequest`
runner, where it would through an SSL invalid error for our
self-generated certificates. Other system approaches of accessing the
same localhost address (ie. in curl) worked fine.

This was due to Chromium using a separate certificate storage location on Linux/Ubuntu
and not respecting the system store. This commit installs a helper
utility in the docker image to manage this custom certificate manager,
and adds the necessary hooks during proxy setup time to install the
certificate in this location as well. We also refactor credentials to
live within a sub-directory of the proxy directories so we can override
the code in docker-compose without overriding the SSL certificates that
have already been installed into the docker system.