Add dialer definition
Introduce a new notion of a `DialerDefinition,` which is a
self-contained route to the open Internet. There are 3rd party proxy-based
dialers (that replace the previous notion of an end proxy) as well as
local dialers (that connect directly from the groove machine to the
internet). Dialers can also specify request conditions. If
these are specified, we will only consider dialers valid for specific
request situations. This is particularly useful in situations like
passthrough where we want to limit based on resource type but might be
more generically relevant to other definition types as well.

To leverage this for MITM transactions we also re-write a custom
RoundTrip handler that progressively tries lower priority dialers
until we have success or reach a failure condition.

`